BIG CHANGES
The main problem I had was that I had not properly set up my character with the camera and certain parts of code were confusing and not efficient. I have re done big parts of the project. but now the player can walk, has a camera following it, and now has tiles where each one is associated with an image, so that I am able to create a map and see the characters movements.